+++ source/index.js
@@ -15,14 +15,16 @@ const createReduxEntry = () => {
   }
 
   const middleware = (reduxMiddlewareStore) => { // this store only gives us `getState` and `dispatch`, no `subscribe`
+    if (typeof (reduxMiddlewareStore.getState) !== 'function' || typeof (reduxMiddlewareStore.dispatch) !== 'function') throw new Error('[ReduxEntry][middleware] invalid reduxMiddlewareStore')
+    if (store !== null) throw new Error('[ReduxEntry][middleware] already set reduxMiddlewareStore')
     store = reduxMiddlewareStore
     return (next) => (action) => (onAction(action) === true) || next(action) // TODO: is this useful? if onAction return true, skip following middleware
   }
 
   const setEntry = (actionType, entryFunction) => {
-    if (typeof (actionType) !== 'string') throw new Error('[ReduxEntry] non-string actionType:', actionType, entryFunction)
-    if (typeof (entryFunction) !== 'function') throw new Error('[ReduxEntry] non-function entryFunction:', actionType, entryFunction)
-    if (entryMap[ actionType ]) console.warn('[ReduxEntry] possible unexpected entry overwrite:', actionType, entryFunction)
+    if (typeof (actionType) !== 'string') throw new Error(`[ReduxEntry][setEntry] non-string actionType: ${actionType}`)
+    if (typeof (entryFunction) !== 'function') throw new Error(`[ReduxEntry][setEntry] non-function entryFunction: ${actionType}, ${entryFunction}`)
+    if (entryMap[ actionType ]) console.warn(`[ReduxEntry][setEntry] possible unexpected entry overwrite: ${actionType}`)
     entryMap[ actionType ] = entryFunction
   }
 
@@ -32,16 +34,18 @@ const createReduxEntry = () => {
 }
 
 // Simple store for state, with a handy function to wrapEntry
-const createStateStore = (state = null) => ({
-  getState: () => state,
-  setState: (nextState) => (state = { ...state, ...nextState }),
-  wrapEntry: (func) => (store, action) => func(state, store, action)
-})
+const createStateStore = (state) => {
+  if (state === undefined) throw new Error('[ReduxEntry][createStateStore] initialState expected')
+  return {
+    getState: () => state,
+    setState: (nextState) => (state = nextState),
+    wrapEntry: (func) => (store, action) => func(state, store, action)
+  }
+}
 
 // Handy function to update both the stateStore and ReduxStore
 const createStateStoreReducer = (actionType, { getState, setState }) => (state, { type, payload }) => { // the reducer, NOTE the action should be like `{ type, payload }`
-  if (type !== actionType) return state
-  setState(payload)
+  type === actionType && setState(payload)
   return getState()
 }